+++ packages/blockera/php/tests/Providers/AssetsProviderTest.php
@@ -7,7 +7,7 @@ use Blockera\Setup\Providers\AppServiceProvider;
 use Blockera\Setup\Providers\EditorAssetsProvider;
 use Illuminate\Contracts\Container\BindingResolutionException;
 
-class TestAssetsProvider extends \Blockera\Dev\PHPUnit\AppTestCase {
+class AssetsProviderTest extends \Blockera\Dev\PHPUnit\AppTestCase {
 
 	protected static ?EditorAssetsProvider $provider = null;
 
@@ -40,11 +40,11 @@ class TestAssetsProvider extends \Blockera\Dev\PHPUnit\AppTestCase {
 
 		if ( false !== strpos( $asset, '-styles' ) ) {
 
-			$this->assertTrue( wp_style_is( '@blockera/' . $asset ) );
+			$this->assertTrue( wp_style_is( '@blockera/' . $asset ), "'$asset' Style is not enqueued" );
 
 		} else {
 
-			$this->assertTrue( wp_script_is( '@blockera/' . $asset ) );
+			$this->assertTrue( wp_script_is( '@blockera/' . $asset ), "'$asset' Script is not enqueued" );
 		}
 	}